feat: unify networkservices api, fuzzer and mapper
