+++ apis/networkservices/v1alpha1/types.generated.go
@@ -12,4 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// +generated:types
+// krm.group: networkservices.cnrm.cloud.google.com
+// krm.version: v1alpha1
+// proto.service: google.cloud.networkservices.v1
+// resource: NetworkServicesServiceBinding:ServiceBinding
+
 package v1alpha1
